remove column comment
