+++ features/republican_silence.feature
@@ -43,7 +43,6 @@ Feature:
     Then I should see "En raison du silence républicain, votre espace est momentanément désactivé. Vous pourrez de nouveau y accéder à la fin de celui-ci."
     Examples:
       | uri                                                             |
-      | /projets-citoyens/13003-en-marche-projet-citoyen/discussions    |
       | /projets-citoyens/13003-en-marche-projet-citoyen/actions/creer  |
 
   @javascript
