--- src/CitizenProject/CitizenProjectCommentCommand.php
@@ -1,70 +1,0 @@
-<?php
-
-namespace AppBundle\CitizenProject;
-
-use AppBundle\Entity\Adherent;
-use AppBundle\Entity\CitizenProject;
-use Symfony\Component\Validator\Constraints as Assert;
-
-class CitizenProjectCommentCommand
-{
-    /**
-     * @var CitizenProject
-     *
-     * @Assert\NotNull
-     */
-    private $citizenProject;
-
-    /**
-     * @var Adherent Any member of the citizen project
-     */
-    private $author;
-
-    /**
-     * @var string|null
-     *
-     * @Assert\NotBlank
-     */
-    private $content;
-
-    /**
-     * @var bool
-     */
-    private $sendMail;
-
-    public function __construct(CitizenProject $citizenProject, ?Adherent $author)
-    {
-        $this->citizenProject = $citizenProject;
-        $this->author = $author;
-    }
-
-    public function getCitizenProject(): CitizenProject
-    {
-        return $this->citizenProject;
-    }
-
-    public function getAuthor(): ?Adherent
-    {
-        return $this->author;
-    }
-
-    public function getContent(): ?string
-    {
-        return $this->content;
-    }
-
-    public function setContent(?string $content): void
-    {
-        $this->content = $content;
-    }
-
-    public function isSendMail(): ?bool
-    {
-        return $this->sendMail;
-    }
-
-    public function setSendMail(?bool $sendMail): void
-    {
-        $this->sendMail = $sendMail;
-    }
-}
