+++ src/Entity/CoordinatorAreaInterface.php
@@ -7,10 +7,6 @@ interface CoordinatorAreaInterface
     public const PRE_APPROVED = 'PRE_APPROVED';
     public const PRE_REFUSED = 'PRE_REFUSED';
 
-    public function getCoordinatorComment(): ?string;
-
-    public function setCoordinatorComment(?string $comment): void;
-
     public function getCreator(): ?Adherent;
 
     public function setCreator(?Adherent $adherent): void;
