--- templates/citizen_project/show_comments.html.twig
@@ -1,33 +1,0 @@
-{% extends 'citizen_project/show_base.html.twig' %}
-
-{% block page_title citizen_project.name %}
-{% block canonical_url url('app_citizen_project_show_comments', {slug: citizen_project.slug}) %}
-
-{% block citizen_project_show_content %}
-    {{ parent() }}
-    {% if can_see_comment_citizen_project(citizen_project) %}
-        <div class="citizen-project-comments">
-            {% if can_comment_citizen_project(citizen_project) %}
-                {{ form_start(form) }}
-                    {{ form_errors(form.content) }}
-                    {{ form_widget(form.content, { attr: { placeholder: "Ecrivez votre message ici (soyez courtois et respectueux)" } }) }}
-                    {{ form_rest(form) }}
-                    <button class="newbtn--small btn--blue">Publier</button>
-                {{ form_end(form) }}
-
-            {% endif %}
-
-            {% for comment in comments %}
-                <div class="citizen-project-comment">
-                    <div class="header">
-                        <p>{{ comment.authorFullName }} </p>
-                        <small> {{ comment.createdAt|ago|capitalize }}</small>
-                    </div>
-                    <div>
-                        <p>{{ comment.content }}</p>
-                    </div>
-                </div>
-            {% endfor %}
-        </div>
-    {% endif %}
-{% endblock %}
